chore: Support React Native v0.63.
Upgrade `react-native` and `react` in `devDependencies`, and also
take the new Flow version used by the new `react-native`. Follow
react-native@87370c4c0 in removing the `inexact-spread` rule, which
doesn't work with the new Flow.

Also, raise the upper bound for `react-native` under
`peerDependencies` to signal to consumers that RN v0.63 is
supported.

No further adjustments seem necessary to ensure compatibility, and
the ones made here are tiny and unlikely to be responsible for
breaking compatibility with versions in the ">=0.57 <=0.62" range,
so narrowing that range isn't indicated to be necessary.

Also, update the example app to work with RN v0.63. Helped by the
example of PR #177, start from a fresh project created with
`react-native init --version=0.63.4`, then add back in the contents
and the particular config for the example app. Maintain that the
changes to the RN boilerplate all have comments with "added for"
near them, so they're easy to find.