chore: Apply Flow's `implicit-inexact-object` lint.
By making the inexact object types explicitly inexact, instead of
implicitly.

This aligns the Flow types with the `flow-typed` project's standard
for writing third-party library definitions [1] for Flow v0.104
(this project uses v0.122, as of daa6a9fae) [2]:

> Any flow-typed definition that is for flow 0.104+ will use only
> explicit exact object types `{||}` or explicit inexact object
> types `{...}`.

This code isn't a libdef hosted by the `flow-typed` project. But the
`flow-typed` project is a good benchmark for how a package's
consumers will expect it to declare its Flow types, if it chooses to
do so.

This change has the benefit of allowing consumers to continue
migrating to `exact_by_default=true` [1] without surprising changes
to this package's types, which would stem from the `{}` syntax being
interpreted as exact instead of inexact. (I see that this project
doesn't currently intend `{}` to mean exact, because it doesn't set
`exact_by_default=true` in its own Flow config.)

[1] https://flow.org/en/docs/libdefs/
[2] https://medium.com/flow-type/how-to-upgrade-to-exact-by-default-object-type-syntax-7aa44b4d08ab